[handlers] Introduce EntryData typed dict
